Added functionality to delete and archive buttons
Resolves #256

As a developer, I added functionality to the delete and archive buttons along with confirmation dialogues. When products are deleted or archived from the frontend, they are hidden. They are marked as isHidden (deleted) or isSold (archived) in the database.

I wanted to make the page reroute back to the category page, but the components complicates it and my attempts to incorporate router.back() caused errors.