belindas-closet-nextjs_8_467_minimized-card (#468)
Resolves #467

This PR improves the appearance of the All Products page by minimizing the information included for each product. It also makes the columns more narrow to allow for more space for other products.

I noticed that there are 4 columns now but there was was a weird space on the right side, so I adjusted the column size to reduce that extra space. I made these changes for category-page too to keep it consistent.