belindas-closet-nextjs_9_475-home-mission-donation-page-responsive-ui
Resolves #475

This PR cleans up the UI for mobile screens by centering the Home page logo above the Sign Up and Sign In buttons properly (the logo was off to the side before). I also added margin above the headers on the Donation and Mission page because they were very close under the nav bar.

I noticed an unused CSS file was made for Donation page, so I deleted it because we are switching to Material UI.