belindas-closet-nextjs_9_478_signup-signin-responsive-pages-responsive-ui
Resolves #478

This PR ensures that the Sign Up and Sign In page has responsive UI and adjusts to mobile screen sizes (and tablet screens) well, with no cutoffs or the need to scroll horizontally. I made additional UI changes to make sure the mascot is resting on top of the Sign Up/Sign In box as closely as possible.

Test this PR by right clicking the page and using Inspect elements to enter the Toggle Device Toolbar and adjust the screen sizes.