belindas-closet-nextjs_9_485_forgot-change-pw-responsive-ui
Resolves #485

This PR cleans up the UI for mobile screens for the Forgot Password and Change Password Pages and overall appearance in desktop screens as well to keep them consistent with other pages on the app.

Test this PR by right clicking the page and using Inspect elements to enter the Toggle Device Toolbar and adjust the screen sizes.