belindas-closet-nextjs_9_488_category-page-all-products-archived-products-pages-ui
Resolves #488

This PR ensures that certain elements, such as the "Edit" and "Archive" buttons, and product detail text are properly centered.

I also tweaked margin spaces and sizes of certain elements so that the space was distributed better (image being large was causing the product detail text to be cut off and pushed to the side).

Smallest mobile screen size: