belindas-closet-nextjs_9_480_remove-unrelated-edit-product-dialog (#484)
* belindas-closet-nextjs_9_480_remove-unrelated-edit-product-dialog

Resolves #480

This PR removes unrelated fields from the Edit Product dialog on the Product Detail page. I also removed the ability to change the product's type because it doesn't make sense to be able to change it. The "Save changes" button is disabled until changes are made to the product.

Pants product:

Shoes product:

* Commit