belindas-closet-nextjs_9_496_product-detail-dialogs-responsive-ui
Resolves #496

WIP draft

This PR ensures that the dialogs on the Product Detail page have responsive UI and adjusts to mobile screen sizes (and tablet screens) well, with no cutoffs or the need to scroll horizontally.

Test this PR by right clicking the page and using Inspect elements to enter the Toggle Device Toolbar and adjust the screen sizes.