belindas-closet-nextjs-500-admin-creator-page-responsive-ui
Resolves #500

This PR ensures that the Admin and Creator pages have responsive UI and adjusts to mobile screen sizes (and tablet screens) well, with no cutoffs or the need to scroll horizontally.

Test this PR by right clicking the page and using Inspect elements to enter the Toggle Device Toolbar and adjust the screen sizes.