belindas-closet-nextjs-10_515_buttons-missing-home-page
Resolves #515

This PR fixes the problem of the Sign Up and Sign In buttons not appearing on the Home page while signing out.

Recreate this bug on the main branch by going to the Home page while signed in, and then sign out while still on the page. The buttons do not appear as they should.